comment_frame classified into new controller
