+++ public/index.php
@@ -17,6 +17,7 @@ $router->add('user_closed', ['controller' => 'UserClosedController', 'action' =>
 $router->add('delete_post', ['controller' => 'DeletePostController', 'action' => 'postDelete']);
 $router->add('ajax_load', ['controller' => 'AjaxLoadPostController', 'action' => 'loadPostAjax']);
 $router->add('confirm_post', ['controller' => 'ConfirmPostController', 'action' => 'confirmPost']);
+$router->add('comment_frame', ['controller' => 'CommentController', 'action' => 'frameComment']);
 $router->add('post', ['controller' => 'PostsController', 'action' => 'loadPostsFriends']);
 $router->add('submit', ['controller' => 'PostsController', 'action' => 'submitPost']);
 $router->add('logged_out', ['controller' => 'LogoutController', 'action' => 'logout']);
